Cria regra para renderização AMP no Share
